refactor: extract execution-log.ts from run.ts (#447)
Step 3: Extract bridge/API helpers, execution logging, cooldown tracking,
and event emission into dedicated module.
